requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.3.0
langchain-community>=0.3.0
duckdb>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.0.0
langgraph-cli